Move ActionTypes to actionTypes, refs #7
