+++ app/actions/uiActions.js
@@ -1,6 +1,6 @@
 import { createAction } from 'redux-actions';
 
-import types from './ActionTypes';
+import types from './actionTypes';
 
 const changeMessage = createAction(types.CHANGE_MESSAGE);
